Initial commit for renaming Execution Manager component to Template Manager
