+++ components/analytics-io/org.wso2.carbon.analytics.eventsink.template.deployer/src/main/resources/META-INF/component.xml
@@ -18,7 +18,7 @@
     <ManagementPermissions>
         <ManagementPermission>
             <DisplayName>Event sink template deploy</DisplayName>
-            <ResourceId>/permission/admin/manage/execution-manager</ResourceId>
+            <ResourceId>/permission/admin/manage/template-manager</ResourceId>
         </ManagementPermission>
     </ManagementPermissions>er
