+++ components/analytics-processors/org.wso2.carbon.analytics.spark.template.deployer/pom.xml
@@ -36,7 +36,7 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.analytics-common</groupId>
-            <artifactId>org.wso2.carbon.event.execution.manager.core</artifactId>
+            <artifactId>org.wso2.carbon.event.template.manager.core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon</groupId>
@@ -92,7 +92,7 @@
                               !org.wso2.carbon.analytics.spark.template.deployer.internal.*,
                           </Export-Package>
                           <Import-Package>
-                              org.wso2.carbon.event.execution.manager.core,
+                              org.wso2.carbon.event.template.manager.core,
                               org.wso2.carbon.analytics.spark.core.*;version="${carbon.analytics.version}",
                               *;resolution:=optional,
                           </Import-Package>
